LA stuff added to /dev
