+++ dev/awk.sh
@@ -1,0 +1,3 @@
+for FILE in *.json; do
+  awk '{ gsub("oldstr","newstr"); print }' ${FILE} > ./output/${FILE}
+done
